@@ -2,10 +2,10 @@ package net.loginbuddy.demoserver.provider;
 
 import net.loginbuddy.common.util.Jwt;
 
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServlet;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
 public class LoginbuddyProviderJwks extends HttpServlet {
