@@ -14,8 +14,8 @@ import net.loginbuddy.common.util.Jwt;
 import net.loginbuddy.common.util.Pkce;
 import org.json.simple.JSONObject;
 
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.util.UUID;
 import java.util.logging.Logger;
