@@ -8,8 +8,8 @@ import net.loginbuddy.common.util.ParameterValidatorResult.RESULT;
 import net.loginbuddy.oidcdr.OIDCDRMaster;
 import org.json.simple.JSONObject;
 
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.util.logging.Logger;
 
