@@ -9,8 +9,8 @@ import net.loginbuddy.config.loginbuddy.Providers;
 import net.loginbuddy.service.util.SessionContext;
 import org.json.simple.JSONObject;
 
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import java.util.logging.Logger;
 
 public class CallbackHandlerImplicit extends Callback implements CallbackHandler {
