@@ -6,9 +6,9 @@ import net.loginbuddy.config.management.AccessToken;
 import net.loginbuddy.config.management.AccessTokenLocation;
 import net.loginbuddy.config.management.ConfigurationTypes;
 
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
