@@ -10,9 +10,9 @@ import net.loginbuddy.config.management.RuntimeTypes;
 import net.loginbuddy.config.scope.LoginbuddyScope;
 import net.loginbuddy.config.scope.RequireScope;
 
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.util.List;
 
