@@ -9,9 +9,9 @@ import net.loginbuddy.common.util.ParameterValidatorResult;
 import net.loginbuddy.service.server.Overlord;
 import org.json.simple.JSONObject;
 
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
 public class Userinfo extends Overlord {
