@@ -3,9 +3,9 @@ package net.loginbuddy.service.server;
 import net.loginbuddy.common.api.HttpHelper;
 import net.loginbuddy.common.util.Jwt;
 
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
 public class Jwks extends Overlord {
