@@ -9,10 +9,10 @@ import net.loginbuddy.common.util.ParameterValidatorResult;
 import net.loginbuddy.sidecar.util.SessionContext;
 import org.jose4j.lang.JoseException;
 
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServlet;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.util.logging.Logger;
 
