@@ -14,9 +14,9 @@ import net.loginbuddy.common.util.ParameterValidatorResult;
 import net.loginbuddy.config.HeadOfInitialize;
 import net.loginbuddy.sidecar.util.SessionContext;
 
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
+import jakarta.servlet.http.HttpServlet;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.util.logging.Logger;
 
